my_teams_backend/requirements.txt
django
djangorestframework
uwsgi
django-cors-headers